+++ src/components/carousel/slide/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./slide.component";
-export type { SlideProps } from "./slide.component";